move ViolationFixer to common
