+++ verilog/analysis/BUILD
@@ -229,13 +229,11 @@ cc_library(
         "//common/analysis:token_stream_linter",
         "//common/analysis:violation_handler",
         "//common/strings:line_column_map",
-        "//common/strings:diff",
         "//common/text:concrete_syntax_tree",
         "//common/text:text_structure",
         "//common/text:token_info",
         "//common/util:file_util",
         "//common/util:logging",
-        "//common/util:user_interaction",
         "//verilog/parser:verilog_token_classifications",
         "//verilog/parser:verilog_token_enum",
         "@com_google_absl//absl/flags:flag",
