+++ verilog/tools/lint/verilog_lint.cc
@@ -189,10 +189,10 @@ int main(int argc, char** argv) {
               << autofix_mode << std::endl;
   }
 
-  const verilog::ViolationFixer::AnswerChooser applyAllFixes =
+  const verible::ViolationFixer::AnswerChooser applyAllFixes =
       [](const verible::LintViolation&,
-         absl::string_view) -> verilog::ViolationFixer::Answer {
-    return {verilog::ViolationFixer::AnswerChoice::kApplyAll, 0};
+         absl::string_view) -> verible::ViolationFixer::Answer {
+    return {verible::ViolationFixer::AnswerChoice::kApplyAll, 0};
   };
 
   auto select_printer =
@@ -211,19 +211,19 @@ int main(int argc, char** argv) {
     case AutofixMode::kPatchInteractive:
       CHECK(autofix_output_stream);
       violation_handler.reset(
-          new verilog::ViolationFixer(&std::cerr, autofix_output_stream));
+          new verible::ViolationFixer(&std::cerr, autofix_output_stream));
       break;
     case AutofixMode::kPatch:
       CHECK(autofix_output_stream);
-      violation_handler.reset(new verilog::ViolationFixer(
+      violation_handler.reset(new verible::ViolationFixer(
           &std::cerr, autofix_output_stream, applyAllFixes));
       break;
     case AutofixMode::kInplaceInteractive:
-      violation_handler.reset(new verilog::ViolationFixer(&std::cerr, nullptr));
+      violation_handler.reset(new verible::ViolationFixer(&std::cerr, nullptr));
       break;
     case AutofixMode::kInplace:
       violation_handler.reset(
-          new verilog::ViolationFixer(&std::cerr, nullptr, applyAllFixes));
+          new verible::ViolationFixer(&std::cerr, nullptr, applyAllFixes));
       break;
   }
